Add-query_params-parameter-to-login-redirect-route
On login for the deployed site Okta does not postpend the URL query string it is supposed to add in the case of a successful log which causes the redirect page to not redirect to the dashboard. In an attempt to solve this issue I am testing if explicitly adding a parameter to the React Router route for the login redirect page's Route component allows that to be added. This has to be testing against the deployed site as the local site does not suffer from the issue.